ref(etl): Move the property test runner into test_utils
Extract the budget/seed property runner from the value roundtrip test
module into etl::test_utils::property so property suites in other
crates (for example destination value roundtrips) can reuse it instead
of copying the harness.

proptest moves from a dev-dependency to an optional dependency enabled
by the test-utils feature, and the module is gated on the feature alone
because plain cfg(test) builds do not enable optional dependencies. The
ROUNDTRIP_PROPERTY_BUDGET_SECS and ROUNDTRIP_PROPERTY_SEED knobs are
unchanged.
